fix(user feedback): Pass event id to feedback dialog in JS examples (#873)
Examples of using the user feedback dialog in other languages pass the event id to the method, but the JS examples (both vanilla and React) don't. Sometimes, race conditions mean that a different event id is associated with the feedback than that of the event itself. Passing the event id prevents that.

This makes that change, and also adds user feedback to the Angular usage example.