Update: Nest Discover beneath Performance and remove deprecated Discover docs (#1549)
This update does two things:

Discover v1 is now deprecated and the docs are out-of-date. This PR removes the deprecated Discover v1 docs.

Discover v2 is up-to-date and is considered a feature of Performance. This PR nests Discover v2 beneath Performance AND renames the doc Discover (instead of Discover v2).