feat(envelopes): Document envelopes (#1640)
Adds documentation for the `Envelope` format and API endpoint that new
SDKs are supposed to use for event ingestion. Along with this,
everything that is specific to the older `/store/` endpoint is moved
into a separate document. Additionally, this also links to the Minidump
and UE4 endpoints, which are documented in their respective platform
docs.