fix: Render markdown formatting in description (#2530)
In this PR, I made the gatsby-plugin-openapi create child nodes for each API path's parameters and description. For this child nodes, I tagged as text/markdown so that the gatsby-plugin-mdx and gatsby-transformer-remark automatically parse the strings. Then we can render the result in the <Markdown> component. This removes the need to dangerouslySetInnerHTML for the params and the description.